requirements.txt
numpy
pandas
kagglehub
python-dotenv